+++ packages/aws-amplify-angular/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.5-unstable.2"></a>
+## [2.0.5-unstable.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.5-unstable.1...aws-amplify-angular@2.0.5-unstable.2) (2018-09-20)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.5-unstable.1"></a>
+## [2.0.5-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.5-unstable.0...aws-amplify-angular@2.0.5-unstable.1) (2018-09-17)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.5-unstable.0"></a>
+## [2.0.5-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.4...aws-amplify-angular@2.0.5-unstable.0) (2018-09-17)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.4"></a>
+## [2.0.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.3...aws-amplify-angular@2.0.4) (2018-09-17)
+
+
+### Bug Fixes
+
+* **aws-amplify-angular:** authState visibility issues on initial load ([#1657](https://github.com/aws-amplify/amplify-js/issues/1657)) ([50611fe](https://github.com/aws-amplify/amplify-js/commit/50611fe))
+
+
+
+
+<a name="2.0.3"></a>
+## [2.0.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.2...aws-amplify-angular@2.0.3) (2018-09-12)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.2"></a>
+## [2.0.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.2-unstable.0...aws-amplify-angular@2.0.2) (2018-09-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.2-unstable.0"></a>
+## [2.0.2-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.1...aws-amplify-angular@2.0.2-unstable.0) (2018-09-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.1-unstable.2...aws-amplify-angular@2.0.1) (2018-09-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.1-unstable.2"></a>
+## [2.0.1-unstable.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.0...aws-amplify-angular@2.0.1-unstable.2) (2018-09-06)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.1-unstable.1"></a>
+## [2.0.1-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.0...aws-amplify-angular@2.0.1-unstable.1) (2018-08-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@1.0.4-unstable.3...aws-amplify-angular@2.0.0) (2018-08-28)
+
+
+### Features
+
+* UI Components ([1ff1abd](https://github.com/aws-amplify/amplify-js/commit/1ff1abd))
+
+
+### BREAKING CHANGES
+
+* UI Components
+
+
+
+
+<a name="1.0.4-unstable.3"></a>
+## [1.0.4-unstable.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@1.0.4-unstable.2...aws-amplify-angular@1.0.4-unstable.3) (2018-08-28)
+
+
+* Amplify ui migration (#1517) ([41d3184](https://github.com/aws-amplify/amplify-js/commit/41d3184)), closes [#1517](https://github.com/aws-amplify/amplify-js/issues/1517)
+
+
+### BREAKING CHANGES
+
+* UI Components
+
+
+
+
 <a name="1.0.4-unstable.2"></a>
 ## [1.0.4-unstable.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@1.0.4-unstable.1...aws-amplify-angular@1.0.4-unstable.2) (2018-08-23)